requirements.txt
streamlit==1.32.0
openai==1.30.0
groq==0.9.0
firebase-admin==6.5.0
streamlit-google-auth==0.1.3
python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.0
pydantic==2.7.0